Move stuff from .h to .cpp files. Add missing headers.
